// src/types/product.type.ts
interface Product {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  tags: Array<string>;
  brand: string;
  sku: string;
  weight: number;
  dimensions: Dimensions;
  warrantyInformation: string;
  shippingInformation: string;
  availabilityStatus: string;
  reviews: Array<Review>;
  returnPolicy: string;
  minimumOrderQuantity: number;
  meta: Meta;
  images: Array<string>;
  thumbnail: string;
  userAdded?: boolean;
}

interface Dimensions {
  width: number;
  height: number;
  depth: number;
}

interface Meta {
  createdAt: string;
  updatedAt: string;
  barcode: string;
  qrCode: string;
}

export interface Review {
  rating: number;
  comment: string;
  date: string;
  reviewerName: string;
  reviewerEmail: string;
}

interface ProductResponse {
  products: Array<Product>;
  total: number;
  skip: number;
  limit: number;
}

export type { Product, ProductResponse };